+++ api/src/main/scala/dx/api/DxWorkflow.scala
@@ -37,7 +37,8 @@ case class DxWorkflowDescribe(project: String,
                               tags: Option[Set[String]] = None,
                               types: Option[Vector[String]] = None,
                               inputs: Option[Vector[IOParameter]] = None,
-                              outputs: Option[Vector[IOParameter]] = None)
+                              outputs: Option[Vector[IOParameter]] = None,
+                              hidden: Option[Boolean] = None)
     extends DxObjectDescribe
 
 object DxWorkflowDescribe {
@@ -114,7 +115,7 @@ case class DxWorkflow(id: String, project: Option[DxProject])(dxApi: DxApi = DxA
     val summary = descFields.get("summary").flatMap(unwrapString)
     val title = descFields.get("title").flatMap(unwrapString)
     val types = descFields.get("types").flatMap(unwrapStringArray)
-    val tags = descFields.get("tags").flatMap(unwrapStringArray).map(_.toSet)
+    val tags = descFields.get("tags").map(DxObject.parseJsonTags)
     val inputSpec = descFields.get("inputSpec") match {
       case Some(JsArray(inps)) => Some(IOParameter.parseIOSpec(dxApi, inps))
       case _                   => None
@@ -131,6 +132,7 @@ case class DxWorkflow(id: String, project: Option[DxProject])(dxApi: DxApi = DxA
       case Some(JsArray(outs)) => Some(IOParameter.parseIOSpec(dxApi, outs))
       case _                   => None
     }
+    val hidden = descFields.get("hidden").flatMap(unwrapBoolean)
     desc.copy(
         inputSpec = inputSpec,
         outputSpec = outputSpec,
@@ -143,7 +145,8 @@ case class DxWorkflow(id: String, project: Option[DxProject])(dxApi: DxApi = DxA
         types = types,
         tags = tags,
         inputs = inputs,
-        outputs = outputs
+        outputs = outputs,
+        hidden = hidden
     )
   }
 
@@ -158,6 +161,13 @@ case class DxWorkflow(id: String, project: Option[DxProject])(dxApi: DxApi = DxA
     jsValue match {
       case JsArray(array) => Some(array.flatMap(unwrapString))
       case _              => None
+    }
+  }
+
+  def unwrapBoolean(jsValue: JsValue): Option[Boolean] = {
+    jsValue match {
+      case JsBoolean(value) => Some(value)
+      case _                => None
     }
   }